Add markercut flag for get_fixation_LT functions
Be default the `markercut` is set to TRUE, which allows more accurate reports for looking time.
That means, if participants were already looking at a certain a AOI before the AOI
before the actual marker onset was defined/appears the old (i.e., `markercut = FALSE) was to add
this pre-release time as well. This can lead to situations where participants would longer look
to a stimuli than it is actually presented. Therefore, `markercut = TRUE` is enabled by default.