Resolves #12
The current strategy for multiple hit_names follows this rules:
(1) If there is only one hit name defined (e.g., one "screen" AOI),
it will continue execution as there is no conflicting other hit names
defined. In other words, the FixationIndex will therefore contain only:
hit_name(1), FALSE (in case the AOI was not hit), and NA (in the case
of no Data). That means, in the further process it will just take the
GazeEventDuration of the one hit_name. There is no conflict.

However, if there two or more hit names within the current fixation
index, the current index will be skipped and the next index will be
processed (the problemeatic index will be looged in
bad_fixation_indexes)