Add get_lookaway_scope_end function
This function returns an updated scope$end vector for a given, df, scope,
and lookaway_stop time criterion in ms.

Whenever a non-fixation event is greater or equal the provided lookaway_stop,
the scope_end will be updated the prior fixation_index.

For example;
This comes in handy if you want to stop calculating looking durations for
subjects when they look away from the screen for the provided time in ms.